requirements.txt
Django>=1.8,<2.0
djangorestframework>=3.6.4
django-webpack-loader==0.5.0
Markdown==2.6.9